Nicolas Regnault: - fix back the RSES in FQHESphereMPSEntanglementSpectrumParticlePartition - fix the component access for vectors > 2^30 in ComplexVector
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@3258 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/HilbertSpace/BosonOnTorusShort.cc
@@ -3257,7 +3257,7 @@ ComplexVector& BosonOnTorusShort::CoreC4Rotation (ComplexVector& inputState, Par
 	  TmpInputSpace->ConvertToMonomial(TmpInputSpace->StateDescription[j], TmpInputSpace->StateKyMax[j] + this->NbrBosons - 1, TmpInputMonomial);
 	  for (int k = 0; k < this->NbrBosons; ++k)
 	    for (int l = 0; l < this->NbrBosons; ++l)
-	      DeterminantMatrix[k][l] = PhaseMatrix[TmpInputMonomial[k]][TmpOutputMonomial[l]];
+	      DeterminantMatrix[k][l] = PhaseMatrix[(int) TmpInputMonomial[k]][(int) TmpOutputMonomial[l]];
 	  TmpLogCoefficient2 = 0.0;
 	  for (int k = 0; k <= TmpInputSpace->TemporaryStateKyMax; ++k)
 	    {
