+++ FQHE/src/HilbertSpace/BosonOnTorusWithMagneticTranslationsShort.cc
@@ -1183,7 +1183,7 @@ ComplexVector& BosonOnTorusWithMagneticTranslationsShort::CoreC4Rotation (Comple
 	  TmpInputSpace->ConvertToMonomial(TmpInputSpace->StateDescription[j], TmpInputSpace->FermionicMaxMomentum, TmpInputMonomial);
 	  for (int k = 0; k < this->NbrBosons; ++k)
 	    for (int l = 0; l < this->NbrBosons; ++l)
-	      DeterminantMatrix[k][l] = PhaseMatrix[TmpInputMonomial[k]][TmpOutputMonomial[l]];
+	      DeterminantMatrix[k][l] = PhaseMatrix[(int) TmpInputMonomial[k]][(int) TmpOutputMonomial[l]];
 	  TmpLogCoefficient2 = 0.0;
 	  for (int k = 0; k <= TmpInputSpace->TemporaryStateKyMax; ++k)
 	    {
